notification_service/requirements.txt
kafka-python
jsonschema
python-dotenv